Release/2.69.9

+++ redisinsight/ui/src/pages/browser/components/delete-key-popover/DeleteKeyPopover.tsx
@@ -1,12 +1,15 @@
-import { EuiButton, EuiButtonIcon, EuiPopover, EuiText } from '@elastic/eui'
-
 import React from 'react'
 
 import cx from 'classnames'
 import { KeyTypes, ModulesKeyTypes } from 'uiSrc/constants'
 import { formatLongName } from 'uiSrc/utils'
 import { RedisResponseBuffer } from 'uiSrc/slices/interfaces'
-import { Spacer } from 'uiSrc/components/base/layout/spacer'
+import {
+  DestructiveButton,
+  IconButton,
+} from 'uiSrc/components/base/forms/buttons'
+import { DeleteIcon } from 'uiSrc/components/base/icons'
+import ConfirmationPopover from 'uiSrc/components/confirmation-popover'
 
 export interface DeleteProps {
   nameString: string
@@ -28,42 +31,43 @@ export const DeleteKeyPopover = ({
   deleting,
   onDelete,
   onOpenPopover,
-}: DeleteProps) => (
-  <EuiPopover
-    anchorClassName={cx('showOnHoverKey', { show: deletePopoverId === rowId })}
-    anchorPosition="leftUp"
-    isOpen={deletePopoverId === rowId}
-    closePopover={() => onOpenPopover(-1, type)}
-    panelPaddingSize="l"
-    button={
-      <EuiButtonIcon
-        iconType="trash"
-        onClick={() => onOpenPopover(rowId, type)}
-        aria-label="Delete Key"
-        data-testid={`delete-key-btn-${nameString}`}
-      />
-    }
-    onClick={(e) => e.stopPropagation()}
-  >
-    <>
-      <EuiText size="m">
-        <h4 style={{ wordBreak: 'break-all' }}>
-          <b>{formatLongName(nameString)}</b>
-        </h4>
-        <EuiText size="s">will be deleted.</EuiText>
-      </EuiText>
-      <Spacer size="m" />
-      <EuiButton
-        fill
-        size="s"
-        color="warning"
-        iconType="trash"
-        isDisabled={deleting}
-        onClick={() => onDelete(name)}
-        data-testid="submit-delete-key"
-      >
-        Delete
-      </EuiButton>
-    </>
-  </EuiPopover>
-)
+}: DeleteProps) => {
+  const onClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    e.stopPropagation()
+    onOpenPopover(rowId, type)
+  }
+
+  return (
+    <ConfirmationPopover
+      anchorClassName={cx('showOnHoverKey', {
+        show: deletePopoverId === rowId,
+      })}
+      anchorPosition="leftCenter"
+      isOpen={deletePopoverId === rowId}
+      closePopover={() => onOpenPopover(-1, type)}
+      panelPaddingSize="l"
+      button={
+        <IconButton
+          icon={DeleteIcon}
+          onClick={onClick}
+          aria-label="Delete Key"
+          data-testid={`delete-key-btn-${nameString}`}
+        />
+      }
+      onClick={(e) => e.stopPropagation()}
+      title={formatLongName(nameString)}
+      message="will be deleted."
+      confirmButton={
+        <DestructiveButton
+          size="small"
+          icon={DeleteIcon}
+          disabled={deleting}
+          onClick={() => onDelete(name)}
+          data-testid="submit-delete-key"
+        >
+          Delete
+        </DestructiveButton>
+      }
+    />
+  )
+}
